third_party/WebKit/LayoutTests/external/wpt/url/toascii.window-expected.txt

This is a testharness.js-based test.
Found 298 tests; 122 PASS, 176 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Loading data…
PASS aa-- (using URL)
PASS aa-- (using URL.host)
PASS aa-- (using URL.hostname)
PASS aa-- (using <a>)
PASS aa-- (using <a>.host)
PASS aa-- (using <a>.hostname)
PASS aa-- (using <area>)
PASS aa-- (using <area>.host)
PASS aa-- (using <area>.hostname)
FAIL a†-- (using URL) Failed to construct 'URL': Invalid URL
FAIL a†-- (using URL.host) assert_equals: expected "xn--a---kp0a" but got ""
FAIL a†-- (using URL.hostname) assert_equals: expected "xn--a---kp0a" but got ""
FAIL a†-- (using <a>) Failed to construct 'URL': Invalid URL
FAIL a†-- (using <a>.host) assert_equals: expected "xn--a---kp0a" but got ""
FAIL a†-- (using <a>.hostname) assert_equals: expected "xn--a---kp0a" but got ""
FAIL a†-- (using <area>) Failed to construct 'URL': Invalid URL
FAIL a†-- (using <area>.host) assert_equals: expected "xn--a---kp0a" but got ""
FAIL a†-- (using <area>.hostname) assert_equals: expected "xn--a---kp0a" but got ""
PASS ab--c (using URL)
PASS ab--c (using URL.host)
PASS ab--c (using URL.hostname)
PASS ab--c (using <a>)
PASS ab--c (using <a>.host)
PASS ab--c (using <a>.hostname)
PASS ab--c (using <area>)
PASS ab--c (using <area>.host)
PASS ab--c (using <area>.hostname)
PASS -x (using URL)
PASS -x (using URL.host)
PASS -x (using URL.hostname)
PASS -x (using <a>)
PASS -x (using <a>.host)
PASS -x (using <a>.hostname)
PASS -x (using <area>)
PASS -x (using <area>.host)
PASS -x (using <area>.hostname)
FAIL -† (using URL) Failed to construct 'URL': Invalid URL
FAIL -† (using URL.host) assert_equals: expected "xn----xhn" but got ""
FAIL -† (using URL.hostname) assert_equals: expected "xn----xhn" but got ""
FAIL -† (using <a>) Failed to construct 'URL': Invalid URL
FAIL -† (using <a>.host) assert_equals: expected "xn----xhn" but got ""
FAIL -† (using <a>.hostname) assert_equals: expected "xn----xhn" but got ""
FAIL -† (using <area>) Failed to construct 'URL': Invalid URL
FAIL -† (using <area>.host) assert_equals: expected "xn----xhn" but got ""
FAIL -† (using <area>.hostname) assert_equals: expected "xn----xhn" but got ""
PASS -x.xn--nxa (using URL)
PASS -x.xn--nxa (using URL.host)
PASS -x.xn--nxa (using URL.hostname)
PASS -x.xn--nxa (using <a>)
PASS -x.xn--nxa (using <a>.host)
PASS -x.xn--nxa (using <a>.hostname)
PASS -x.xn--nxa (using <area>)
PASS -x.xn--nxa (using <area>.host)
PASS -x.xn--nxa (using <area>.hostname)
FAIL -x.β (using URL) Failed to construct 'URL': Invalid URL
FAIL -x.β (using URL.host) assert_equals: expected "-x.xn--nxa" but got ""
FAIL -x.β (using URL.hostname) assert_equals: expected "-x.xn--nxa" but got ""
FAIL -x.β (using <a>) Failed to construct 'URL': Invalid URL
FAIL -x.β (using <a>.host) assert_equals: expected "-x.xn--nxa" but got ""
FAIL -x.β (using <a>.hostname) assert_equals: expected "-x.xn--nxa" but got ""
FAIL -x.β (using <area>) Failed to construct 'URL': Invalid URL
FAIL -x.β (using <area>.host) assert_equals: expected "-x.xn--nxa" but got ""
FAIL -x.β (using <area>.hostname) assert_equals: expected "-x.xn--nxa" but got ""
PASS x-.xn--nxa (using URL)
PASS x-.xn--nxa (using URL.host)
PASS x-.xn--nxa (using URL.hostname)
PASS x-.xn--nxa (using <a>)
PASS x-.xn--nxa (using <a>.host)
PASS x-.xn--nxa (using <a>.hostname)
PASS x-.xn--nxa (using <area>)
PASS x-.xn--nxa (using <area>.host)
PASS x-.xn--nxa (using <area>.hostname)
FAIL x-.β (using URL) Failed to construct 'URL': Invalid URL
FAIL x-.β (using URL.host) assert_equals: expected "x-.xn--nxa" but got ""
FAIL x-.β (using URL.hostname) assert_equals: expected "x-.xn--nxa" but got ""
FAIL x-.β (using <a>) Failed to construct 'URL': Invalid URL
FAIL x-.β (using <a>.host) assert_equals: expected "x-.xn--nxa" but got ""
FAIL x-.β (using <a>.hostname) assert_equals: expected "x-.xn--nxa" but got ""
FAIL x-.β (using <area>) Failed to construct 'URL': Invalid URL
FAIL x-.β (using <area>.host) assert_equals: expected "x-.xn--nxa" but got ""
FAIL x-.β (using <area>.hostname) assert_equals: expected "x-.xn--nxa" but got ""
PASS x..xn--nxa (using URL)
PASS x..xn--nxa (using URL.host)
PASS x..xn--nxa (using URL.hostname)
PASS x..xn--nxa (using <a>)
PASS x..xn--nxa (using <a>.host)
PASS x..xn--nxa (using <a>.hostname)
PASS x..xn--nxa (using <area>)
PASS x..xn--nxa (using <area>.host)
PASS x..xn--nxa (using <area>.hostname)
FAIL x..β (using URL) Failed to construct 'URL': Invalid URL
FAIL x..β (using URL.host) assert_equals: expected "x..xn--nxa" but got ""
FAIL x..β (using URL.hostname) assert_equals: expected "x..xn--nxa" but got ""
FAIL x..β (using <a>) Failed to construct 'URL': Invalid URL
FAIL x..β (using <a>.host) assert_equals: expected "x..xn--nxa" but got ""
FAIL x..β (using <a>.hostname) assert_equals: expected "x..xn--nxa" but got ""
FAIL x..β (using <area>) Failed to construct 'URL': Invalid URL
FAIL x..β (using <area>.host) assert_equals: expected "x..xn--nxa" but got ""
FAIL x..β (using <area>.hostname) assert_equals: expected "x..xn--nxa" but got ""
FAIL xn--a (using URL) assert_throws: function "() => makeURL("url", hostTest.input)" did not throw
FAIL xn--a (using URL.host) assert_equals: expected "x" but got "xn--a"
FAIL xn--a (using URL.hostname) assert_equals: expected "x" but got "xn--a"
FAIL xn--a (using <a>) assert_equals: expected "" but got "xn--a"
FAIL xn--a (using <a>.host) assert_equals: expected "x" but got "xn--a"
FAIL xn--a (using <a>.hostname) assert_equals: expected "x" but got "xn--a"
FAIL xn--a (using <area>) assert_equals: expected "" but got "xn--a"
FAIL xn--a (using <area>.host) assert_equals: expected "x" but got "xn--a"
FAIL xn--a (using <area>.hostname) assert_equals: expected "x" but got "xn--a"
FAIL xn--a.xn--nxa (using URL) assert_throws: function "() => makeURL("url", hostTest.input)" did not throw
FAIL xn--a.xn--nxa (using URL.host) assert_equals: expected "x" but got "xn--a.xn--nxa"
FAIL xn--a.xn--nxa (using URL.hostname) assert_equals: expected "x" but got "xn--a.xn--nxa"
FAIL xn--a.xn--nxa (using <a>) assert_equals: expected "" but got "xn--a.xn--nxa"
FAIL xn--a.xn--nxa (using <a>.host) assert_equals: expected "x" but got "xn--a.xn--nxa"
FAIL xn--a.xn--nxa (using <a>.hostname) assert_equals: expected "x" but got "xn--a.xn--nxa"
FAIL xn--a.xn--nxa (using <area>) assert_equals: expected "" but got "xn--a.xn--nxa"
FAIL xn--a.xn--nxa (using <area>.host) assert_equals: expected "x" but got "xn--a.xn--nxa"
FAIL xn--a.xn--nxa (using <area>.hostname) assert_equals: expected "x" but got "xn--a.xn--nxa"
PASS xn--a.β (using URL)
FAIL xn--a.β (using URL.host) assert_equals: expected "x" but got ""
FAIL xn--a.β (using URL.hostname) assert_equals: expected "x" but got ""
FAIL xn--a.β (using <a>) assert_equals: expected "https://xn--a.β/x" but got "https://xn--a.%CE%B2/x"
FAIL xn--a.β (using <a>.host) assert_equals: expected "x" but got ""
FAIL xn--a.β (using <a>.hostname) assert_equals: expected "x" but got ""
FAIL xn--a.β (using <area>) assert_equals: expected "https://xn--a.β/x" but got "https://xn--a.%CE%B2/x"
FAIL xn--a.β (using <area>.host) assert_equals: expected "x" but got ""
FAIL xn--a.β (using <area>.hostname) assert_equals: expected "x" but got ""
PASS xn--nxa.xn--nxa (using URL)
PASS xn--nxa.xn--nxa (using URL.host)
PASS xn--nxa.xn--nxa (using URL.hostname)
PASS xn--nxa.xn--nxa (using <a>)
PASS xn--nxa.xn--nxa (using <a>.host)
PASS xn--nxa.xn--nxa (using <a>.hostname)
PASS xn--nxa.xn--nxa (using <area>)
PASS xn--nxa.xn--nxa (using <area>.host)
PASS xn--nxa.xn--nxa (using <area>.hostname)
PASS xn--nxa.β (using URL)
PASS xn--nxa.β (using URL.host)
PASS xn--nxa.β (using URL.hostname)
PASS xn--nxa.β (using <a>)
PASS xn--nxa.β (using <a>.host)
PASS xn--nxa.β (using <a>.hostname)
PASS xn--nxa.β (using <area>)
PASS xn--nxa.β (using <area>.host)
PASS xn--nxa.β (using <area>.hostname)
PASS ab--c.xn--nxa (using URL)
PASS ab--c.xn--nxa (using URL.host)
PASS ab--c.xn--nxa (using URL.hostname)
PASS ab--c.xn--nxa (using <a>)
PASS ab--c.xn--nxa (using <a>.host)
PASS ab--c.xn--nxa (using <a>.hostname)
PASS ab--c.xn--nxa (using <area>)
PASS ab--c.xn--nxa (using <area>.host)
PASS ab--c.xn--nxa (using <area>.hostname)
FAIL ab--c.β (using URL) Failed to construct 'URL': Invalid URL
FAIL ab--c.β (using URL.host) assert_equals: expected "ab--c.xn--nxa" but got ""
FAIL ab--c.β (using URL.hostname) assert_equals: expected "ab--c.xn--nxa" but got ""
FAIL ab--c.β (using <a>) Failed to construct 'URL': Invalid URL
FAIL ab--c.β (using <a>.host) assert_equals: expected "ab--c.xn--nxa" but got ""
FAIL ab--c.β (using <a>.hostname) assert_equals: expected "ab--c.xn--nxa" but got ""
FAIL ab--c.β (using <area>) Failed to construct 'URL': Invalid URL
FAIL ab--c.β (using <area>.host) assert_equals: expected "ab--c.xn--nxa" but got ""
FAIL ab--c.β (using <area>.hostname) assert_equals: expected "ab--c.xn--nxa" but got ""
PASS ‍.example (using URL)
FAIL ‍.example (using URL.host) assert_equals: expected "x" but got ""
FAIL ‍.example (using URL.hostname) assert_equals: expected "x" but got ""
FAIL ‍.example (using <a>) assert_equals: expected "https://‍.example/x" but got "https://%E2%80%8D.example/x"
FAIL ‍.example (using <a>.host) assert_equals: expected "x" but got ""
FAIL ‍.example (using <a>.hostname) assert_equals: expected "x" but got ""
FAIL ‍.example (using <area>) assert_equals: expected "https://‍.example/x" but got "https://%E2%80%8D.example/x"
FAIL ‍.example (using <area>.host) assert_equals: expected "x" but got ""
FAIL ‍.example (using <area>.hostname) assert_equals: expected "x" but got ""
FAIL xn--1ug.example (using URL) assert_throws: function "() => makeURL("url", hostTest.input)" did not throw
FAIL xn--1ug.example (using URL.host) assert_equals: expected "x" but got "xn--1ug.example"
FAIL xn--1ug.example (using URL.hostname) assert_equals: expected "x" but got "xn--1ug.example"
FAIL xn--1ug.example (using <a>) assert_equals: expected "" but got "xn--1ug.example"
FAIL xn--1ug.example (using <a>.host) assert_equals: expected "x" but got "xn--1ug.example"
FAIL xn--1ug.example (using <a>.hostname) assert_equals: expected "x" but got "xn--1ug.example"
FAIL xn--1ug.example (using <area>) assert_equals: expected "" but got "xn--1ug.example"
FAIL xn--1ug.example (using <area>.host) assert_equals: expected "x" but got "xn--1ug.example"
FAIL xn--1ug.example (using <area>.hostname) assert_equals: expected "x" but got "xn--1ug.example"
PASS يa (using URL)
FAIL يa (using URL.host) assert_equals: expected "x" but got ""
FAIL يa (using URL.hostname) assert_equals: expected "x" but got ""
FAIL يa (using <a>) assert_equals: expected "https://يa/x" but got "https://%D9%8Aa/x"
FAIL يa (using <a>.host) assert_equals: expected "x" but got ""
FAIL يa (using <a>.hostname) assert_equals: expected "x" but got ""
FAIL يa (using <area>) assert_equals: expected "https://يa/x" but got "https://%D9%8Aa/x"
FAIL يa (using <area>.host) assert_equals: expected "x" but got ""
FAIL يa (using <area>.hostname) assert_equals: expected "x" but got ""
FAIL xn--a-yoc (using URL) assert_throws: function "() => makeURL("url", hostTest.input)" did not throw
FAIL xn--a-yoc (using URL.host) assert_equals: expected "x" but got "xn--a-yoc"
FAIL xn--a-yoc (using URL.hostname) assert_equals: expected "x" but got "xn--a-yoc"
FAIL xn--a-yoc (using <a>) assert_equals: expected "" but got "xn--a-yoc"
FAIL xn--a-yoc (using <a>.host) assert_equals: expected "x" but got "xn--a-yoc"
FAIL xn--a-yoc (using <a>.hostname) assert_equals: expected "x" but got "xn--a-yoc"
FAIL xn--a-yoc (using <area>) assert_equals: expected "" but got "xn--a-yoc"
FAIL xn--a-yoc (using <area>.host) assert_equals: expected "x" but got "xn--a-yoc"
FAIL xn--a-yoc (using <area>.hostname) assert_equals: expected "x" but got "xn--a-yoc"
FAIL ශ්‍රී (using URL) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using URL.host) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using URL.hostname) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using <a>) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using <a>.host) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using <a>.hostname) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using <area>) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using <area>.host) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL ශ්‍රී (using <area>.hostname) assert_equals: expected "xn--10cl1a0b660p" but got "xn--10cl1a0b"
FAIL نامه‌ای (using URL) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using URL.host) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using URL.hostname) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using <a>) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using <a>.host) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using <a>.hostname) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using <area>) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using <area>.host) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
FAIL نامه‌ای (using <area>.hostname) assert_equals: expected "xn--mgba3gch31f060k" but got "xn--mgba3gch31f"
PASS �.com (using URL)
FAIL �.com (using URL.host) assert_equals: expected "x" but got ""
FAIL �.com (using URL.hostname) assert_equals: expected "x" but got ""
FAIL �.com (using <a>) assert_equals: expected "https://\ufffd.com/x" but got "https://%EF%BF%BD.com/x"
FAIL �.com (using <a>.host) assert_equals: expected "x" but got ""
FAIL �.com (using <a>.hostname) assert_equals: expected "x" but got ""
FAIL �.com (using <area>) assert_equals: expected "https://\ufffd.com/x" but got "https://%EF%BF%BD.com/x"
FAIL �.com (using <area>.host) assert_equals: expected "x" but got ""
FAIL �.com (using <area>.hostname) assert_equals: expected "x" but got ""
FAIL xn--zn7c.com (using URL) assert_throws: function "() => makeURL("url", hostTest.input)" did not throw
FAIL xn--zn7c.com (using URL.host) assert_equals: expected "x" but got "xn--zn7c.com"
FAIL xn--zn7c.com (using URL.hostname) assert_equals: expected "x" but got "xn--zn7c.com"
FAIL xn--zn7c.com (using <a>) assert_equals: expected "" but got "xn--zn7c.com"
FAIL xn--zn7c.com (using <a>.host) assert_equals: expected "x" but got "xn--zn7c.com"
FAIL xn--zn7c.com (using <a>.hostname) assert_equals: expected "x" but got "xn--zn7c.com"
FAIL xn--zn7c.com (using <area>) assert_equals: expected "" but got "xn--zn7c.com"
FAIL xn--zn7c.com (using <area>.host) assert_equals: expected "x" but got "xn--zn7c.com"
FAIL xn--zn7c.com (using <area>.hostname) assert_equals: expected "x" but got "xn--zn7c.com"
PASS x01234567890123456789012345678901234567890123456789012345678901x (using URL)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using URL.host)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using URL.hostname)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using <a>)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using <a>.host)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using <a>.hostname)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using <area>)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using <area>.host)
PASS x01234567890123456789012345678901234567890123456789012345678901x (using <area>.hostname)
FAIL x01234567890123456789012345678901234567890123456789012345678901† (using URL) Failed to construct 'URL': Invalid URL
FAIL x01234567890123456789012345678901234567890123456789012345678901† (using URL.host) assert_equals: expected "xn--x01234567890123456789012345678901234567890123456789012345678901-6963b" but got ""
FAIL x01234567890123456789012345678901234567890123456789012345678901† (using URL.hostname) assert_equals: expected "xn--x01234567890123456789012345678901234567890123456789012345678901-6963b" but got ""
FAIL x01234567890123456789012345678901234567890123456789012345678901† (using <a>) Failed to construct 'URL': Invalid URL
FAIL x01234567890123456789012345678901234567890123456789012345678901† (using <a>.host) assert_equals: expected "xn--x01234567890123456789012345678901234567890123456789012345678901-6963b" but got ""
FAIL x01234567890123456789012345678901234567890123456789012345678901† (using <a>.hostname) assert_equals: expected "xn--x01234567890123456789012345678901234567890123456789012345678901-6963b" but got ""
FAIL x01234567890123456789012345678901234567890123456789012345678901† (using <area>) Failed to construct 'URL': Invalid URL
FAIL x01234567890123456789012345678901234567890123456789012345678901† (using <area>.host) assert_equals: expected "xn--x01234567890123456789012345678901234567890123456789012345678901-6963b" but got ""
FAIL x01234567890123456789012345678901234567890123456789012345678901† (using <area>.hostname) assert_equals: expected "xn--x01234567890123456789012345678901234567890123456789012345678901-6963b" but got ""
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa (using URL)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa (using URL.host)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa (using URL.hostname)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa (using <a>)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa (using <a>.host)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa (using <a>.hostname)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa (using <area>)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa (using <area>.host)
PASS x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa (using <area>.hostname)
FAIL x01234567890123456789012345678901234567890123456789012345678901x.β (using URL) Failed to construct 'URL': Invalid URL
FAIL x01234567890123456789012345678901234567890123456789012345678901x.β (using URL.host) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa" but got ""
FAIL x01234567890123456789012345678901234567890123456789012345678901x.β (using URL.hostname) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa" but got ""
FAIL x01234567890123456789012345678901234567890123456789012345678901x.β (using <a>) Failed to construct 'URL': Invalid URL
FAIL x01234567890123456789012345678901234567890123456789012345678901x.β (using <a>.host) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa" but got ""
FAIL x01234567890123456789012345678901234567890123456789012345678901x.β (using <a>.hostname) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa" but got ""
FAIL x01234567890123456789012345678901234567890123456789012345678901x.β (using <area>) Failed to construct 'URL': Invalid URL
FAIL x01234567890123456789012345678901234567890123456789012345678901x.β (using <area>.host) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa" but got ""
FAIL x01234567890123456789012345678901234567890123456789012345678901x.β (using <area>.hostname) assert_equals: expected "x01234567890123456789012345678901234567890123456789012345678901x.xn--nxa" but got ""
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using URL)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using URL.host)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using URL.hostname)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using <a>)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using <a>.host)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using <a>.hostname)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using <area>)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using <area>.host)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.x (using <area>.hostname)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa (using URL)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa (using URL.host)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa (using URL.hostname)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa (using <a>)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa (using <a>.host)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa (using <a>.hostname)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa (using <area>)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa (using <area>.host)
PASS 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa (using <area>.hostname)
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.β (using URL) Failed to construct 'URL': Invalid URL
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.β (using URL.host) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa" but got ""
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.β (using URL.hostname) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa" but got ""
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.β (using <a>) Failed to construct 'URL': Invalid URL
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.β (using <a>.host) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa" but got ""
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.β (using <a>.hostname) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa" but got ""
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.β (using <area>) Failed to construct 'URL': Invalid URL
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.β (using <area>.host) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa" but got ""
FAIL 01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.β (using <area>.hostname) assert_equals: expected "01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.01234567890123456789012345678901234567890123456789.0123456789012345678901234567890123456789012345678.xn--nxa" but got ""
Harness: the test ran to completion.
